requirements.txt
pypdf
pdfplumber
PyMuPDF
pdf2image
torch
pytesseract
numpy
opencv-python
Pillow
ultralytics